chore: dedupe path translate/simplify and exit-edit i18n
