+++ apps/web/src/i18n/locales/en.ts
@@ -844,7 +844,6 @@ const en = {
     pathEditAddAnchor: 'Add anchor',
     pathEditCurve: 'Curve',
     pathEditExit: 'Exit edit',
-    pathEditDone: 'Exit edit (Esc)',
     unitPx: 'px',
     minimap: 'Minimap',
     grid: 'Grid snap',
@@ -1462,7 +1461,6 @@ const en = {
     pencilEasingEaseInExpo: 'Ease in expo',
     pencilEasingEaseOutExpo: 'Ease out expo',
     pencilEasingEaseInOutExpo: 'Ease in-out expo',
-    exitPenEdit: 'Exit edit',
     appearance: 'Appearance',
     multiSelected: '{{count}} objects selected',
     multiSelectedHint: 'Move, delete (Delete), or reorder layers ([ / ]) together. Shift-click to add or remove from the selection.',
