+++ apps/web/src/i18n/locales/ja.ts
@@ -822,7 +822,6 @@ const ja = {
     pathEditAddAnchor: 'アンカーを追加',
     pathEditCurve: '曲線',
     pathEditExit: '編集を終了',
-    pathEditDone: '編集を終了 (Esc)',
     unitPx: 'px',
     minimap: 'ミニマップ',
     grid: 'グリッド吸着',
@@ -1344,7 +1343,6 @@ const ja = {
     pencilEasingEaseInExpo: '指数イン',
     pencilEasingEaseOutExpo: '指数アウト',
     pencilEasingEaseInOutExpo: '指数インアウト',
-    exitPenEdit: '編集を終了',
     appearance: '外観',
     multiSelected: '{{count}} 個のオブジェクトを選択中',
     multiSelectedHint: 'まとめて移動・削除（Delete）、レイヤー順序の変更（[ / ]）ができます。Shift+クリックで選択を追加/解除。',
