+++ apps/web/src/i18n/locales/zh-CN.ts
@@ -836,7 +836,6 @@ const zhCN = {
     pathEditAddAnchor: '增加锚点',
     pathEditCurve: '曲线',
     pathEditExit: '退出编辑',
-    pathEditDone: '退出编辑 (Esc)',
     unitPx: '像素',
     minimap: '小地图',
     grid: '网格吸附',
@@ -1454,7 +1453,6 @@ const zhCN = {
     pencilEasingEaseInExpo: '指数缓入',
     pencilEasingEaseOutExpo: '指数缓出',
     pencilEasingEaseInOutExpo: '指数缓入缓出',
-    exitPenEdit: '退出编辑',
     appearance: '外观',
     multiSelected: '已选中 {{count}} 个对象',
     multiSelectedHint: '可一起拖动、删除（Delete），或调整图层顺序（[ / ]）。按住 Shift 点击可增减选中。',
