+++ apps/web/src/i18n/locales/zh-TW.ts
@@ -816,7 +816,6 @@ const zhTW = {
     pathEditAddAnchor: '增加錨點',
     pathEditCurve: '曲線',
     pathEditExit: '退出編輯',
-    pathEditDone: '退出編輯 (Esc)',
     unitPx: '像素',
     minimap: '小地圖',
     grid: '網格吸附',
@@ -1338,7 +1337,6 @@ const zhTW = {
     pencilEasingEaseInExpo: '指數緩入',
     pencilEasingEaseOutExpo: '指數緩出',
     pencilEasingEaseInOutExpo: '指數緩入緩出',
-    exitPenEdit: '退出編輯',
     appearance: '外觀',
     multiSelected: '已選中 {{count}} 個物件',
     multiSelectedHint: '可一起拖曳、刪除（Delete），或調整圖層順序（[ / ]）。按住 Shift 點擊可增減選取。',
